Move AzCopy to commands folder
